+++ src/components/common/InputComment/InputEditedComment/InputEditedComment.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { ChangeEvent, SetStateAction, KeyboardEvent } from 'react';
 import { css } from '@emotion/react';
-import Button from '@/components/design/Button/Button';
+import Button from '@/components/common/Button/Button';
 import Input from '@/components/design/Input/Input';
 import { useEditComment } from '@/hooks/comment/useEditComment';
