+++ src/components/common/InputsUserInfo/InputCode.tsx
@@ -3,7 +3,7 @@ import { MemberForm } from '@/types/member';
 import { css } from '@emotion/react';
 import React, { ChangeEvent, useEffect } from 'react';
 import { useCheckCode } from '../../../hooks/common/inputsUserInfo/useCheckCode';
-import Button from '../../design/Button/Button';
+import Button from '../Button/Button';
 import Input from '../../design/Input/Input';
 
 interface InputCodeProps {
