+++ src/components/common/InputsUserInfo/InputEmail.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { ChangeEvent, useEffect } from 'react';
 import { css } from '@emotion/react';
-import Button from '../../design/Button/Button';
+import Button from '../Button/Button';
 import Input from '../../design/Input/Input';
 import { useCheckEmail } from '../../../hooks/common/inputsUserInfo/useCheckEmail';
