+++ src/components/common/InputsUserInfo/InputProfileImage/InputProfileImage.tsx
@@ -6,7 +6,7 @@ import React, { useCallback, useState } from 'react';
 import { useDropzone } from 'react-dropzone';
 import defaultProfileImage from '../../../../assets/images/defaultProfile.png';
 import { Theme } from '../../../../styles/Theme';
-import Button from '../../../design/Button/Button';
+import Button from '../../Button/Button';
 import Label from '../../../design/Label/Label';
 import ProfileImage from '../../Profile/ProfileImage/ProfileImage';
 import {
