+++ src/components/common/Modal/ChangeVoteModal/ChangeVoteModal.tsx
@@ -1,5 +1,5 @@
 import React, { SetStateAction, useEffect, useRef } from 'react';
-import Button from '@/components/design/Button/Button';
+import Button from '@/components/common/Button/Button';
 import { useMemberQuery } from '@/hooks/api/useMemberQuery';
 import { useParseJwt } from '@/hooks/common/useParseJwt';
 import { useNewSelector } from '@/store';
