+++ src/pages/LoginPage/sections/LoginForm/LoginForm.tsx
@@ -2,7 +2,7 @@ import React, { useRef } from 'react';
 import { css } from '@emotion/react';
 import Input from '../../../../components/design/Input/Input';
 import { useLoginForm } from '../../../../hooks/login/useLoginForm';
-import Button from '../../../../components/design/Button/Button';
+import Button from '../../../../components/common/Button/Button';
 import { btnLogin, inputContainer } from './LoginForm.style';
 import { Email, Key } from '../../../../assets';
