+++ src/pages/MyPage/UpdatePage/UpdatePage.tsx
@@ -8,7 +8,7 @@ import InputNickname from '@/components/common/InputsUserInfo/InputNickname';
 import InputProfileImage from '@/components/common/InputsUserInfo/InputProfileImage/InputProfileImage';
 import InputPw from '@/components/common/InputsUserInfo/InputPw';
 import InputPwCheck from '@/components/common/InputsUserInfo/InputPwCheck';
-import Button from '@/components/design/Button/Button';
+import Button from '@/components/common/Button/Button';
 import Heading from '@/components/design/Heading/Heading';
 import Input from '@/components/design/Input/Input';
 import { useMemberUpdate } from '@/hooks/mypage/memberUpdate/useMemberUpdate';
